+++ magma/lib/magma/query/question_format.rb
